handle 64-bit

+++ L1/src/L1.h
@@ -27,7 +27,7 @@ namespace L1 {
   struct Item {
     int type;         // defined by L1::ITEM_*TYPE*
     std::string name; // for register & label
-    int value;        // for register (offset) & number
+    int64_t value;        // for register (offset) & number
   };
 
   struct Instruction {
